1.6.1
Extensive update to atask.
Literally every is now customizable.
Specify hooks before and after tasks, replace counters with
\dynamicaufgabencounter{} and \dynamicteilaufgabencounter{}, as well as
set a static counter with \staticaufgabencounter
(\staticteilaufgabencounter, respectively), "Task" and "Subtask" are now
treated as prefixes and may be replaced with \aufgabenprefix as well as
\teilaufgabenprefix. There also is a suffix after the optional title you
may set by \aufgabensuffix, \teilaufgabensuffix. Other than that, you
can customize the margins for tasks and subtasks. For more information,
see the examples.